refactor: extract header state hook
